+++ tests/unit/models/test_blanket_library.py
@@ -9,6 +9,9 @@ from process.data_structure import (
     fwbs_variables,
     physics_variables,
 )
+from process.data_structure import build_variables as bv
+from process.data_structure import fwbs_variables as fwbs
+from process.data_structure import physics_variables as pv
 from process.models.blankets.blanket_library import BlanketLibrary
 from process.models.fw import FirstWall
 
@@ -779,7 +782,6 @@ class LiquidBreederPropertiesParam(NamedTuple):
     dr_blkt_inboard: Any = None
     dr_blkt_outboard: Any = None
     ip: Any = None
-    ofile: Any = None
     expected_den_liq: Any = None
     expected_specific_heat_liq: Any = None
     expected_thermal_conductivity_liq: Any = None
@@ -815,7 +817,6 @@ class LiquidBreederPropertiesParam(NamedTuple):
             dr_blkt_inboard=0.70000000000000007,
             dr_blkt_outboard=1,
             ip=0,
-            ofile=11,
             expected_den_liq=9753.2497999999996,
             expected_specific_heat_liq=189.12018,
             expected_thermal_conductivity_liq=9.238260167312621,
@@ -854,7 +855,6 @@ class LiquidBreederPropertiesParam(NamedTuple):
             dr_blkt_inboard=0.70000000000000007,
             dr_blkt_outboard=1,
             ip=0,
-            ofile=11,
             expected_den_liq=305.30702851374997,
             expected_specific_heat_liq=34.640761200690406,
             expected_dynamic_viscosity_liq=0.00037298826343426359,
@@ -985,7 +985,6 @@ class PressureDropParam(NamedTuple):
     b_bz_liq: Any = None
     roughness_fw_channel: Any = None
     ip: Any = None
-    ofile: Any = None
     i_ps: Any = None
     num_90: Any = None
     num_180: Any = None
@@ -1008,7 +1007,6 @@ class PressureDropParam(NamedTuple):
             b_bz_liq=0.20000000000000001,
             roughness_fw_channel=9.9999999999999995e-07,
             ip=0,
-            ofile=11,
             i_ps=1,
             num_90=2,
             num_180=0,
@@ -1018,6 +1016,24 @@ class PressureDropParam(NamedTuple):
             vv=32.753134225223164,
             label="Inboard first wall",
             expected_pressure_drop_out=36213.58989742931,
+        ),
+        PressureDropParam(
+            radius_fw_channel=1.0,
+            radius_pipe_90_deg_bend=1.0,
+            radius_pipe_180_deg_bend=1.0,
+            a_bz_liq=1.0,
+            b_bz_liq=1.0,
+            roughness_fw_channel=1e-6,
+            ip=0,
+            i_ps=2,
+            num_90=1.0,
+            num_180=1.0,
+            l_pipe=1.0,
+            den=1.0,
+            vsc=1.0,
+            vv=1.0,
+            label="label",
+            expected_pressure_drop_out=1.4325633520224854,
         ),
     ],
 )
@@ -1068,7 +1084,6 @@ class LiquidBreederPressureDropMhdParam(NamedTuple):
     bz_channel_conduct_liq: Any = None
     th_wall_secondary: Any = None
     ip: Any = None
-    ofile: Any = None
     vel: Any = None
     vsc: Any = None
     conduct_liq: Any = None
@@ -1092,7 +1107,6 @@ class LiquidBreederPressureDropMhdParam(NamedTuple):
             bz_channel_conduct_liq=833000,
             th_wall_secondary=0.012500000000000001,
             ip=0,
-            ofile=11,
             vel=0.061438753831945352,
             vsc=0.0017477589503313536,
             conduct_liq=861800.80431007256,
@@ -1112,7 +1126,6 @@ class LiquidBreederPressureDropMhdParam(NamedTuple):
             bz_channel_conduct_liq=833000,
             th_wall_secondary=0.012500000000000001,
             ip=0,
-            ofile=11,
             vel=0.061438753831945352,
             vsc=0.0017477589503313536,
             conduct_liq=861800.80431007256,
@@ -1132,7 +1145,6 @@ class LiquidBreederPressureDropMhdParam(NamedTuple):
             bz_channel_conduct_liq=833000,
             th_wall_secondary=0.012500000000000001,
             ip=0,
-            ofile=11,
             vel=0.061438753831945352,
             vsc=0.0017477589503313536,
             conduct_liq=861800.80431007256,
@@ -1152,7 +1164,6 @@ class LiquidBreederPressureDropMhdParam(NamedTuple):
             bz_channel_conduct_liq=833000,
             th_wall_secondary=0.012500000000000001,
             ip=0,
-            ofile=11,
             vel=0.042500391943592931,
             vsc=0.0017477589503313536,
             conduct_liq=861800.80431007256,
@@ -1519,3 +1530,157 @@ def test_calculate_elliptical_blkt_volumes(
     assert vol_blkt_total == pytest.approx(
         calculateellipticalblktvolumesparam.expected_vol_blkt_total
     )
+
+
+def test_hydraulic_diameter(monkeypatch, blanket_library_fixture):
+    """
+    Test for hydraulic_diameter function.
+    """
+    # Set var values
+    monkeypatch.setattr(fwbs, "radius_fw_channel", 1.0)
+    monkeypatch.setattr(fwbs, "a_bz_liq", 1.0)
+    monkeypatch.setattr(fwbs, "b_bz_liq", 1.0)
+
+    # hydraulic_diameter input = i_channel_shape: 1 = circle, 2 = rectangle
+    # 2.0D0*radius_fw_channel
+    assert blanket_library_fixture.pipe_hydraulic_diameter(1) == pytest.approx(2.0)
+    # 2*a_bz_liq*b_bz_liq/(a_bz_liq+b_bz_liq)
+    assert blanket_library_fixture.pipe_hydraulic_diameter(2) == pytest.approx(1.0)
+
+
+def test_elbow_coeff(blanket_library_fixture):
+    """
+    Test for elbow_coeff function.
+    """
+    # input = r_elbow, ang_elbow, lambda, dh
+    assert blanket_library_fixture.elbow_coeff(1, 0, 1, 1) == pytest.approx(
+        0.0, rel=1e-3
+    )
+    assert blanket_library_fixture.elbow_coeff(1, 90, 1, 1) == pytest.approx(
+        1.7807963267948965, rel=1e-3
+    )
+    assert blanket_library_fixture.elbow_coeff(1, 180, 1, 1) == pytest.approx(
+        3.291157766597427, rel=1e-3
+    )
+    assert blanket_library_fixture.elbow_coeff(1, 90, 1, 0.1) == pytest.approx(
+        15.774371098812502, rel=1e-3
+    )
+    assert blanket_library_fixture.elbow_coeff(0.1, 90, 1, 1) == pytest.approx(
+        66.57, rel=1e-3
+    )
+    assert blanket_library_fixture.elbow_coeff(1, 90, 0.1, 1) == pytest.approx(
+        0.3670796326794896, rel=1e-3
+    )
+
+
+def test_flow_velocity(monkeypatch, blanket_library_fixture):
+    """
+    Test for flow_velocity function.
+    """
+    # Set var values
+    monkeypatch.setattr(fwbs, "radius_fw_channel", 1.0)
+    monkeypatch.setattr(fwbs, "a_bz_liq", 1.0)
+    monkeypatch.setattr(fwbs, "b_bz_liq", 1.0)
+
+    # input = i_channel_shape, mass_flow_rate, flow_density
+    assert blanket_library_fixture.flow_velocity(1, 1, 1) == pytest.approx(
+        0.318, rel=1e-3
+    )
+    assert blanket_library_fixture.flow_velocity(2, 1, 1) == pytest.approx(1.0)
+    assert blanket_library_fixture.flow_velocity(1, 0, 1) == pytest.approx(0.0)
+    assert blanket_library_fixture.flow_velocity(2, 0, 1) == pytest.approx(0.0)
+
+
+def test_liquid_breeder_properties_part_1(monkeypatch, blanket_library_fixture):
+    """
+    Test for liquid_breeder_properties procedure.
+    PbLi or Li, with inboard blanket, no inlet/outlet temp difference.
+    """
+    # Set var values
+    monkeypatch.setattr(fwbs, "a_bz_liq", 0.2)
+    monkeypatch.setattr(pv, "b_plasma_toroidal_on_axis", 6.0)
+    monkeypatch.setattr(pv, "rmajor", 8.0)
+    monkeypatch.setattr(pv, "aspect", 3.0)
+    monkeypatch.setattr(bv, "dr_blkt_inboard", 0.1)
+    monkeypatch.setattr(bv, "dr_blkt_outboard", 0.2)
+    monkeypatch.setattr(fwbs, "i_blkt_inboard", 1)
+    monkeypatch.setattr(fwbs, "inlet_temp_liq", 1.0)
+    monkeypatch.setattr(fwbs, "outlet_temp_liq", 1.0)
+
+    # PbLi - see [Fer2020] for relavent equations
+    monkeypatch.setattr(fwbs, "i_blkt_liquid_breeder_type", 0)
+
+    blanket_library_fixture.liquid_breeder_properties()
+
+    assert pytest.approx(fwbs.den_liq, rel=1e-3) == 1.052e4
+    assert pytest.approx(fwbs.specific_heat_liq, rel=1e-3) == 195.0
+    assert pytest.approx(fwbs.thermal_conductivity_liq, rel=1e-3) == -3.384
+    assert pytest.approx(fwbs.dynamic_viscosity_liq, rel=1e-3) == 0.0155
+    assert pytest.approx(fwbs.electrical_conductivity_liq, rel=1e-3) == 9.71e5
+
+    assert pytest.approx(fwbs.b_mag_blkt, rel=1e-3) == (9.085, 4.458)
+    assert pytest.approx(fwbs.hartmann_liq, rel=1e-3) == (7.189e3, 3.528e3)
+
+    # Li - see [Lyublinski et al., 2009] for relavent equations
+    monkeypatch.setattr(fwbs, "i_blkt_liquid_breeder_type", 1)
+
+    blanket_library_fixture.liquid_breeder_properties()
+
+    assert pytest.approx(fwbs.den_liq, rel=1e-3) == 504.0
+    assert pytest.approx(fwbs.specific_heat_liq, rel=1e-3) == 2.833e6
+    assert pytest.approx(fwbs.dynamic_viscosity_liq, rel=1e-3) == 1.051e112
+    assert pytest.approx(fwbs.electrical_conductivity_liq, rel=1e-3) == 9.27e8
+    assert pytest.approx(fwbs.b_mag_blkt, rel=1e-3) == (9.085, 4.458)
+    assert pytest.approx(fwbs.hartmann_liq, rel=1e-3) == (2.7e-53, 1.3e-53)
+
+    # con_vsc_rat = electrical_conductivity_liq/dynamic_viscosity_liq
+    # hartmann_liq = b_mag_blkt * a_bz_liq/2.0D0 * sqrt(con_vsc_rat)
+
+
+def test_liquid_breeder_properties_part_2(monkeypatch, blanket_library_fixture):
+    """
+    Test for liquid_breeder_properties procedure. No inboard blanket.
+    """
+    # Set var values
+    monkeypatch.setattr(fwbs, "a_bz_liq", 0.2)
+    monkeypatch.setattr(pv, "b_plasma_toroidal_on_axis", 6.0)
+    monkeypatch.setattr(pv, "rmajor", 8.0)
+    monkeypatch.setattr(pv, "aspect", 3.0)
+    monkeypatch.setattr(bv, "dr_blkt_inboard", 0.0)
+    monkeypatch.setattr(bv, "dr_blkt_outboard", 0.2)
+    monkeypatch.setattr(fwbs, "i_blkt_inboard", 0)
+    monkeypatch.setattr(fwbs, "i_blkt_liquid_breeder_type", 0)
+    monkeypatch.setattr(fwbs, "inlet_temp_liq", 0.0)
+    monkeypatch.setattr(fwbs, "outlet_temp_liq", 0.0)
+
+    blanket_library_fixture.liquid_breeder_properties()
+
+    assert pytest.approx(fwbs.b_mag_blkt, rel=1e-3) == (8.999, 4.458)
+
+
+def test_liquid_breeder_properties_part_3(monkeypatch, blanket_library_fixture):
+    """
+    Test for liquid_breeder_properties procedure.
+    With inlet/outlet temp difference.
+    """
+    # Set var values
+    monkeypatch.setattr(fwbs, "a_bz_liq", 0.2)
+    monkeypatch.setattr(pv, "b_plasma_toroidal_on_axis", 6.0)
+    monkeypatch.setattr(pv, "rmajor", 8.0)
+    monkeypatch.setattr(pv, "aspect", 3.0)
+    monkeypatch.setattr(bv, "dr_blkt_inboard", 0.1)
+    monkeypatch.setattr(bv, "dr_blkt_outboard", 0.2)
+    monkeypatch.setattr(fwbs, "i_blkt_inboard", 1)
+    monkeypatch.setattr(fwbs, "inlet_temp_liq", 0.0)
+    monkeypatch.setattr(fwbs, "outlet_temp_liq", 1.0)
+
+    # PbLi - see [Fer2020] for relavent equations
+    monkeypatch.setattr(fwbs, "i_blkt_liquid_breeder_type", 0)
+
+    blanket_library_fixture.liquid_breeder_properties()
+    assert pytest.approx(fwbs.den_liq, rel=1e-3) == 1.052e4
+    # Li - see [Lyublinski et al., 2009] for relavent equations
+    monkeypatch.setattr(fwbs, "i_blkt_liquid_breeder_type", 1)
+
+    blanket_library_fixture.liquid_breeder_properties()
+    assert pytest.approx(fwbs.den_liq, rel=1e-3) == 504.0
